Add Transaction service and controller layers

+++ src/main/java/az/spring/response/AccountResponse.java
@@ -5,13 +5,15 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class AccountResponse {
 
     private String accountNumber;
     private Boolean status;
-    private Double balance;
+    private BigDecimal balance;
 
     @JsonIgnore
     private User user;
